requirements.txt
django==1.7.3
paho-mqtt==1.1